Rename pkg/plugin/manager -> pkg/installer
