+++ pkg/installer/README.md
