+++ pkg/installer/RanchergMSACredentialProvider.dll
