+++ pkg/installer/install-plugin.ps1
