+++ pkg/installer/uninstall-plugin.ps1
