+++ pkg/installer/upgrade_windows.go
@@ -1,4 +1,4 @@
-package manager
+package installer
 
 import (
 	"bytes"
